Rename/move `Contribute_With_Google` to `Contribute_With_Google_Block` for consistency with SiWG.
